requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Image Processing
Pillow==10.1.0

# Database & Caching
redis==5.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Job Queue
rq==1.15.1

# Utilities
python-magic==0.4.27
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0

# Data handling
pyyaml==6.0.1
